+++ src/maicos/modules/densitycylinder.py
@@ -9,6 +9,9 @@
 """Module for computing cylindrical density profiles."""
 
 import logging
+from typing import List, Optional, Union
+
+import MDAnalysis as mda
 
 from ..core import ProfileCylinderBase
 from ..lib.util import render_docs
@@ -39,21 +42,21 @@ class DensityCylinder(ProfileCylinderBase):
 
     def __init__(
         self,
-        atomgroups,
-        dens="mass",
-        dim=2,
-        zmin=None,
-        zmax=None,
-        bin_width=1,
-        rmin=0,
-        rmax=None,
-        refgroup=None,
-        grouping="atoms",
-        unwrap=True,
-        bin_method="com",
-        output="density.dat",
-        concfreq=0,
-        jitter=0.0,
+        atomgroups: Union[mda.AtomGroup, List[mda.AtomGroup]],
+        dens: str = "mass",
+        dim: int = 2,
+        zmin: Optional[float] = None,
+        zmax: Optional[float] = None,
+        bin_width: float = 1,
+        rmin: float = 0,
+        rmax: Optional[float] = None,
+        refgroup: Optional[mda.AtomGroup] = None,
+        grouping: str = "atoms",
+        unwrap: bool = True,
+        bin_method: str = "com",
+        output: str = "density.dat",
+        concfreq: float = 0,
+        jitter: float = 0.0,
     ):
         super().__init__(
             weighting_function=density_weights,