+++ CHANGELOG.rst
@@ -15,9 +15,11 @@ v0.8 (XXXX/XX/XX)
 -----------------
 Simon Gravelle, Philip Loche, Marc Sauter
 
+- Add typehints for modules (!253)
 - Only test minimum and maximum Python version in CI (!252)
 - Add typehints for core classes (!251)
 - Update documentation with parameters, returns and examples for library functions
+  (!248)
 - Update CI to use latest MacOS (!250)
 - Add tables to documentation pages (!249)
 - Fix links to own classes in examples (!247)